# [combine][perf] Use more efficient way to compute dividers.