… usecases/vectorizer/combine_test.go