### What's being changed:
It's cheaper to keep track of dividers using `int32` instead of `float32`. To get an idea about the difference I've added a benchmark. Results are below:
before:
```
Running tool: /opt/homebrew/bin/go test -benchmem -run=^$ -bench ^BenchmarkCombine$ github.com/weaviate/weaviate/usecases/vectorizer

goos: darwin
goarch: arm64
pkg: github.com/weaviate/weaviate/usecases/vectorizer
BenchmarkCombine-8      14267766            80.67 ns/op       64 B/op          2 allocs/op
PASS
ok      github.com/weaviate/weaviate/usecases/vectorizer    1.369s
```
after:
```
Running tool: /opt/homebrew/bin/go test -benchmem -run=^$ -bench ^BenchmarkCombine$ github.com/weaviate/weaviate/usecases/vectorizer

goos: darwin
goarch: arm64
pkg: github.com/weaviate/weaviate/usecases/vectorizer
BenchmarkCombine-8      17251676            68.93 ns/op       64 B/op          2 allocs/op
PASS
ok      github.com/weaviate/weaviate/usecases/vectorizer    2.382s
```

So for relatively small vectors the speedup is ~15% without any downsides.